Set bio/en data to be integers. Math.ceil didn't do that automatically.
git-svn-id: svn://svn.code.sf.net/p/ospace/code/branches/ospace_0.5@207 9ec6d687-074c-4f17-8882-0daec8a15dc2

+++ server/lib/ige/ospace/IPlanet.py
@@ -483,8 +483,8 @@ class IPlanet(IObject):
 			maxPop = int(maxPop)
 			# eat
 			pop = obj.storPop / 1000.0
-			wantBio = math.ceil(pop * obj.popEatBio)
-			wantEn = math.ceil(pop * obj.popEatEn)
+			wantBio = int(math.ceil(pop * obj.popEatBio))
+			wantEn = int(math.ceil(pop * obj.popEatEn))
 			# auto regulation of min resources
 			if obj.autoMinStor:
 				obj.minBio += wantBio * Rules.autoMinStorTurns
